+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 
 ### Fixed
 - Blurring of headings in popover hints with text hiding enabled.
+- Width of small buttons container in locked skill popout boxes.
 
 ### Changed
 - Custom html attributes in options page now have "data-" prefix.